+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+  - admin: new filters for list of tokens
+  - admin: new filters for list of stable tokens
+  - admin: actions to list/de-list tokens
+
+### Changed
+
+  - erc20: tokens are not listed by default. Hub Operators need to list them through the admin
+
+### Removed
+
+  - erc20 (API): users are no longer able to add new tokens (potential security risk)
+
+
 ## [0.4.2] - 2022-03-25
 
 ### Added